Update left- and fully-anchored types for more consistency
New anchor types are set up to use some of the basic
character substitutions (2/two) and the standard tokenization
routine. These are both only useful as phrases (either quoted
or in the `pf` (e)dismax argument.

This includes both the updates to schema.xml and the inclusion
of the new .jar file

Other minor changes:
  * Change definition of punctuation to include control characters,
    math symbols, and currency symbols
  * Remmove ordinals->numbers (e.g., "two" -> "second") from synonyms.txt.